requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
supabase==2.9.1
watchdog==4.0.1
python-multipart==0.0.9
PyJWT[crypto]==2.9.0
anthropic==0.52.0
openai==1.82.0
